Refactor, bundler, fix build
