+++ cardanow-ts/test/snapshot.test.ts
@@ -1,5 +1,5 @@
 import { expect, test, describe } from 'vitest'
-import { SnapshotExporter, SnapshotConfig } from '../snapshot'
+import { SnapshotExporter, SnapshotConfig } from '../src/snapshot'
 
 
 const alwaysSuccessConfig: SnapshotConfig = {
